requirements.txt
duckdb>=1.5.1
altair>=5.3.0
pandas==2.2.3
requests==2.32.3
python-dotenv==1.0.1
groq>=0.9.0
langchain==0.3.19
langchain-groq==0.3.7
streamlit==1.43.2
scipy==1.15.2
numpy==2.2.3
apscheduler==3.11.0
pytest==8.3.5
ruff==0.9.9
tenacity==9.0.0
